unify settingsListTiles
- make settings a separate lib
