+++ src/ui/flutter_app/lib/shared/settings/src/card.dart
@@ -16,7 +16,7 @@
   along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-import 'package:flutter/material.dart';
+part of '../settings.dart';
 
 class SettingsCard extends StatelessWidget {
   const SettingsCard({
